Fixed problem where the 'fixture offset' (which is expressed as 'workplane()' in the python scripts) was not being constrained between values 1 and 9 inclusive.
The 'fixture offset' value was also not being read in from the XML file using the same attribute
name that was used to write it.  Hence it was always defaulting to a value of '0'.

Fixed the Python code that sets the coordinate system (G54 -> G59.3) based on the fixture offset
found in the DepthOp operation.