Added cutting_edge_height parameter to the CuttingTool object.  This is the length of the cutting surface from the bottom of the tool.
Initialised the 'stepdown' parameter to match the cutting tool's cutting_edge_height if
the AdaptiveRoughing object referred to a cuttingTool object as well.  This seems
aggressive but I believe the idea behind the Adaptive Clearing algorithm is that it
improves tool life by cutting with the whole cutting edge rather than just the bottom
bit.  I think the idea is to edge into the material with much smaller sideways cuts
to maintain chip load.

Added some warning dialogs to indicate if the machines.txt file is not setup correctly.  The
side-effect of this is a failure in both the python and gcode generation steps.  Again,
warnings were added at all three points to help explain why things aren't working.