Added RS274X (gerber) import function.  It's a VERY long way from finished but it's at a point where it serves my purposes.  It imports line and circle elements but it doesn't (yet) offset those lines to become wide sketches suitable for isolation routing using profile operations.
The existing version has been useful to me to allow me to scale the SVG file that I imported from Inkscape.  i.e. when I import a raster image of the board layout into Inkscape, I lose scale.  I have been able to use this import tool to import some of the graphics from the gerber data and scale the imported SVG graphics to that scale.

I should also mention that the gerber data is that generated by the gEDA PCB application.  It's not a completely implemented RS274X import routine.  (and neither do I intend to make it so - others, feel free if you wish).