Reverted change to NCCode when Dan pointed out that GCode doesn't do rotated arcs.  Instead, it produces a helical arc whose axis is still the 'current' axis but where the Z values move progressively between the start and end Z values.
Changed GCode generator for Contour operations so that, if the fixture has a rotation in either the XZ or YZ planes, the rotated arcs are converted into smaller lines that follow the rotated arc.   If there is no rotation in the XZ or YZ planes then proper arc movements are generated.