Another one of those really big check-ins.
This adds the 'private fixture' functionality asked for in Issue 189. i.e. you can right click on a machine operation object (Contour, Pocket etc.) and add a fixture to that operation alone.  Each time it adds another fixture to another machine operation, it assigns the next fixture number.  If you then go and update the second fixture's coordinate number (in its properties), the code looks for a pre-existing fixture of that same number and merges the two.  i.e. both operations can share the same private fixture object.

It is still possible to add 'public' fixtures in the same way as they have always been.  These 'public fixtures' will operate only on those machine operations that do not have a private fixture amongst their children.

This change also modifies the passing of a CFixture object to all AppendTextToProgram() methods.  Instead, I have defined a CMachineState class.  A pointer to a CMachineState object is passed into (and out of) each AppendTextToProgram() method.  This allows each machine operation to update the machine state in the form of the current fixture, current tool and the current location.  This allows, not only for the private fixtures functionality but for the fact that the Inlay operation changes tools within the single operation.  Eventually, it will allow the sorting of drilling locations to take into account the tool's current location as at the end of the previous machining operation.

Sorry about the big change but I'm pretty sure it will run in alright.  It works find on both Linux and Windows (famous last words).